backend/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.3
dj-database-url==2.3.0
Django==5.2.1
django-allauth==65.8.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
fido2==2.0.0
idna==3.10
oauthlib==3.2.2
pycparser==2.22
PyJWT==2.10.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.4.0
whitenoise==6.9.0
